Serializer: Add basic support for deserializing to generic types
SweetJson's way of deserializing to generic types is different than that
of Gson. We check the signature of the class and associate type args
that match the field's generic type. This is still an experiment and I'm
not sure if this is better than what Gson does. But experiments are
good aren't they? Furthermore, this initial implementation does not
handle type arguments that themselves are parameterized.

These are fine: `List<String>, List<E>, Map<String, String>, E[]`.
These aren't: `List<List<String>>, List<List<E>> etc.`